Remove test gates
